// api-client/models/admin-team-group-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AdminTeamEntryDTO } from './admin-team-entry-dto';
import { GroupCategory } from './group-category';

/**
 * 
 * @export
 * @interface AdminTeamGroupDTO
 */
export interface AdminTeamGroupDTO {
    /**
     * 
     * @type {GroupCategory}
     * @memberof AdminTeamGroupDTO
     */
    'category': GroupCategory;
    /**
     * 
     * @type {Array<AdminTeamEntryDTO>}
     * @memberof AdminTeamGroupDTO
     */
    'entries': Array<AdminTeamEntryDTO>;
}