// api-client/models/group-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export enum GroupCategory {
    VEDENI_SERVER = 'VEDENI_SERVER',
    VEDENI_TEAM = 'VEDENI_TEAM',
    HELPER_TEAM = 'HELPER_TEAM',
    MODERATOR_TEAM = 'MODERATOR_TEAM',
    BUILDER_TEAM = 'BUILDER_TEAM',
    TECHNIK_TEAM = 'TECHNIK_TEAM',
    CHEATHUNTER_TEAM = 'CHEATHUNTER_TEAM',
    YOUTUBER = 'YOUTUBER',
    TIKTOKER = 'TIKTOKER'
}