// api-client/models/media-group-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { GroupCategory } from './group-category';
import { MediaEntryDTO } from './media-entry-dto';

/**
 * 
 * @export
 * @interface MediaGroupDTO
 */
export interface MediaGroupDTO {
    /**
     * 
     * @type {GroupCategory}
     * @memberof MediaGroupDTO
     */
    'category': GroupCategory;
    /**
     * 
     * @type {Array<MediaEntryDTO>}
     * @memberof MediaGroupDTO
     */
    'entries': Array<MediaEntryDTO>;
}